separate commands into their own modules
so adding commands will be easier in the future. also its easier to digest where each command has its own help and execute function in its own module (IMO)
